requirements.txt
Django==3.2.3
django-cors-headers==3.7.0
django-environ==0.8.1
django-filter==2.4.0
django-rest-knox==4.1.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.0.0
psycopg2-binary==2.8.6
whitenoise==5.3.0
gunicorn==20.1.0